refactor: 重构窗口消息系统与组件架构，引入交互系统与 DPI 适配
核心重构
- 重命名 CompGroup -> ComponentGroup，新增 ComponentGroup.hpp/.cpp
- ComponentGroup 添加递归命中检测 getComponentAt()，支持子组件树查找
- Component 添加 contains() 方法，用于组件的点包含测试
- 新增 isNotLogicalGroup() 区分逻辑分组与实体分组

交互系统
- 新增 Interactive 组件基类 (include/comp/Interactive.hpp)
- 新增 InputTypes 定义鼠标按钮枚举 MouseButtonType 与转换函数
- 新增 ComponentEventContext 组件级事件上下文
- 窗口鼠标点击消息自动派发到交互式组件的 onClicking/onClicked 事件

窗口消息系统
- 新增 WindowMessageDispatcher 窗口消息调度器，封装消息分发逻辑
- 新增 WindowEventContext / WindowMessageEventContext 窗口事件上下文
- 窗口过程函数支持 onMessageDispatching / onMessageDispatched 前后事件
- EventContext 新增 cancelAction() / isCanceled() 支持默认操作取消

DPI 适配
- 将 DPI 计算从静态注册移到窗口构造后，使用 GetDpiForWindow() 获取真实 DPI
- DPICHANGED 时同步更新 dpiScale，支持运行时 DPI 变化
- 移除 GetDpiForSystem() 调用，改为窗口实际的 DPI 比例
- 添加 Window::getDpiScale() 方法供子组件使用

事件系统改进
- next() 重命名为 resetForNext()，语义更清晰
- 模板参数 EventArg 重命名为 EventContextArg
- EventContext 各方法添加 noexcept 标注
- trigger() 中每次调用前正确重置上下文状态

渲染器改进
- D2D1_FACTORY_TYPE_MULTI_THREADED -> SINGLE_THREADED
- 构造函数中初始化 DPI，避免后续 DPICHANGED 前使用默认值
- paint() 使用 BeginPaint/EndPaint 正确包裹，修复 WM_PAINT 行为

类型系统改进
- WindowHandle 的 HWND value 改为私有成员，封装性提升
- TextFormat 使用 string_view 替代 string，避免不必要的拷贝
- Position2F 添加 LPARAM 构造函数，支持从窗口消息直接构建
- Util 添加 string_view 版本的 utf8ToWide() 重载

代码质量
- 添加 WIL 错误处理（THROW_LAST_ERROR_IF 替代裸调用）
- RegisterClassEx 添加错误检查
- 添加 WM_NCDESTROY 处理，正确移除窗口子类
- Window 添加析构函数自动调用 release()
- 修复各种 clang-tidy 警告与 narrowing conversion 问题
- 添加 .opencode/skills/cpp-code-review/SKILL.md
- 添加 docs/TODO.md 记录后续开发计划
- 更新 vcpkg.json 依赖配置

+++ src/type/Color.cpp
@@ -1,5 +1,7 @@
 #include "type/Color.hpp"
 
+// NOLINTBEGIN(cppcoreguidelines-avoid-magic-numbers, readability-magic-numbers)
+
 namespace gauzy::type
 {
     Color::Color(const std::uint32_t hex) noexcept : value(hex)
@@ -7,7 +9,7 @@ namespace gauzy::type
     }
 
     Color::Color(const std::uint32_t rgb, const float alpha) noexcept :
-        value( static_cast<std::uint8_t>(alpha * 255) << 24 | rgb)
+        value(static_cast<std::uint8_t>(alpha * 255) << 24 | rgb)
     {
     }
 
@@ -86,6 +88,8 @@ namespace gauzy::type
 
     Color::operator D2D1::ColorF() const noexcept
     {
-        return {rgb(), a() / 255.0f};
+        return { rgb(), static_cast<float>(a()) / 255.0F };
     }
 }
+
+// NOLINTEND(cppcoreguidelines-avoid-magic-numbers, readability-magic-numbers)
